+++ feature/google/src/latestDependencies/java/com/revenuecat/purchases/google/productTypeConversions.kt
@@ -1,5 +1,3 @@
-package com.revenuecat.purchases.google
-
 import com.android.billingclient.api.BillingClient
 import com.revenuecat.purchases.ProductType
