+++ feature/google/src/main/java/com/revenuecat/purchases/google/storeProductConversions.kt
@@ -3,6 +3,7 @@ package com.revenuecat.purchases.google
 import com.android.billingclient.api.SkuDetails
 import com.revenuecat.purchases.models.StoreProduct
 import org.json.JSONObject
+import toProductType
 
 fun SkuDetails.toStoreProduct() =
     StoreProduct(
